+++ src/components/UI/ButtonSubmit/ButtonSubmit.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import './ButtonSubmit.scss'
+const ButtonSubmit = ({buttonTex}) => {
+    return (
+        <div className='ButtonSubmit'>
+            {buttonTex}
+        </div>
+    );
+};
+
+export default ButtonSubmit;